requirements.txt
rumps==0.4.0
psutil==5.9.5
Pillow==10.0.0 